Refactor: move test to dedicated package for better organization
